refactor(cache): extract scm-common-cache module from scm-common-data (P1-1 phase 1)
