+++ scm-common/cache/src/main/java/com/scmcloud/common/cache/metrics/CacheHealthIndicator.java
@@ -1,4 +1,4 @@
-package com.scmcloud.common.web.health;
+package com.scmcloud.common.cache.metrics;
 
 import com.scmcloud.common.cache.spring.TwoLevelCache;
 import com.scmcloud.common.cache.spring.TwoLevelCacheManager;
