+++ scm-common/cache/src/main/java/com/scmcloud/common/redis/lock/LockAcquisitionException.java
