+++ scm-common/data/pom.xml
@@ -21,6 +21,16 @@
     <dependencies>
         <dependency>
             <groupId>com.scmcloud</groupId>
+            <artifactId>scm-common-core</artifactId>
+            <version>1.0.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>com.scmcloud</groupId>
+            <artifactId>scm-common-cache</artifactId>
+            <version>1.0.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>com.scmcloud</groupId>
             <artifactId>scm-common-security-core</artifactId>
             <version>1.0.0-SNAPSHOT</version>
         </dependency>
@@ -89,11 +99,6 @@
             <groupId>com.baomidou</groupId>
             <artifactId>mybatis-plus-jsqlparser</artifactId>
         </dependency>
-        <!-- redis -->
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-data-redis</artifactId>
-        </dependency>
         <!-- database end -->
 
         <!-- tools start -->
@@ -109,12 +114,6 @@
         <dependency>
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-databind</artifactId>
-        </dependency>
-        <!-- caffeine -->
-        <dependency>
-            <groupId>com.github.ben-manes.caffeine</groupId>
-            <artifactId>caffeine</artifactId>
-            <version>3.2.2</version>
         </dependency>
         <dependency>
             <groupId>io.micrometer</groupId>
