+++ scm-common/pom.xml
@@ -19,6 +19,7 @@
 
     <modules>
         <module>core</module>
+        <module>cache</module>
         <module>data</module>
         <module>web</module>
         <module>monitoring</module>
